refactor: encapsulated add/remove/position logic into SceneUpdater class for better modularity and maintainability
